app/requirements.txt
python-telegram-bot
waitress
requests
flask[async]
flask-executor
face_recognition
python-dotenv
Pillow
numpy